Add note for manual discovery on TPlink plugs
For some reason I was not able to auto-discovery my HS105 plug. It was unclear what type I should specify, as I couldn't specify a `host` without a type. I tried `light` first, which resulted in a configuration error, and then moved on to `switch`, which brought the component into the UI.